# Move synchronization out of instrument and into scan.