This is the first chunk of changes. It removes synchronization between scan and cameras from the instrument and uses a new "scan box simulator" that simulates the behavior of hardware connections on real devices.
All instrumentation-kit tests except for one also pass with the patch submitted in instrumentation-kit. The failing test fails on this line: https://github.com/nion-software/nionswift-instrumentation-kit/blob/b2745d38e97f1c459750c0c347c4741d890f439f/nion/instrumentation/test/ScanControl_test.py#L446
I wasn't able to figure out why this fails now...if you add a ``scan_hardware_source.get_next_xdatas_to_start()`` after ``scan_hardware_source.start_playing()`` it passes.