@cmeyer 2022-02-01
See https://github.com/nion-software/nionswift-instrumentation-kit/pull/107#issuecomment-1026333689

I did a quick read over of this - all looks good - but I might have more comments after discussing the comment above.

Also, is the time slice to 50ms necessary? If so, why?