@Brow71189 2022-02-01
> 
> Also, is the time slice to 50ms necessary? If so, why?

This was mostly a test to see if and how the time slice length impacts performance. It doesn't seem to make an impact, so we can change it back. But note that at least on windows the maximum timer resolution is 20 ms, so any sleep shorter than that will be at least 20 ms long (i.e. 5 ms doesn't work).