@cmeyer 2022-02-01
> > Also, is the time slice to 50ms necessary? If so, why?
> 
> This was mostly a test to see if and how the time slice length impacts performance. It doesn't seem to make an impact, so we can change it back. But note that at least on windows the maximum timer resolution is 20 ms, so any sleep shorter than that will be at least 20 ms long (i.e. 5 ms doesn't work).

Let's change it back since it isn't related to this PR. It's useful to evaluate how that delay affects performance separately, though. As I recall, I purposely made it short to highlight issues on macOS - but it may be unnecessarily short now.