@cmeyer 2022-02-02
I did some speed tests on just the scan simulator. The new code is quite a bit slower. Can you take a look and try to understand why it's so much slower?

| Scan Size | OLD latency | NEW latency | OLD fps | NEW fps |
| --- | --- | --- | --- | --- |
| 256px | 7ms | 14ms | 19fps | 19fps |
| 1024px | 9ms | 40ms | 19fps | 12fps |