@cmeyer 2022-02-02
Measuring a basic SI (30px, be sure to play scan before doing SI) - the new code is significantly faster!

| Code |  Size | Exposure | Actual Time | Theoretical Time |
| --- | --- | --- | --- | --- |
| Old | 30 | 5ms | 50s | 4.8s |
| Old | 30 | 20ms | 48s | 19.2s |
| New | 30 | 5ms | *17s* | 4.8s |
| New | 30 | 20ms | *25s* | 19.2s |