@cmeyer 2022-02-07
The timing looks good again with the latest commit (6b43f4e)