@Brow71189 2022-02-08
> The timing looks good again with the latest commit ([6b43f4e](https://github.com/nion-software/nionswift-usim/commit/6b43f4ee8e4c99b374a823bfa25fed41855eba00))

Great, I'd say we can merge this then.